Updates data type for id to a wider type

+++ scottadams/game/model.py
@@ -3,7 +3,7 @@
 
 """
 
-from sqlalchemy import Column, Integer, String
+from sqlalchemy import Column, Integer, BigInteger, String
 from sqlalchemy.ext.declarative import declarative_base
 
 Base = declarative_base()
@@ -14,7 +14,7 @@ class User(Base):
     """
     __tablename__ = 'users'
 
-    sender_id = Column(Integer, primary_key=True)
+    sender_id = Column(BigInteger, primary_key=True)
     game_id = Column(Integer, primary_key=True)
     bitflags = Column(Integer, default=0)
     current_location = Column(Integer)
